fix(nginx): re add curl command in nginx base image
We recently change the base image of our graviteeio/nginx:1.27.5 from nginx:1.27.5-alpine to nginx:1.27.5-alpine-slimto reduce the number of library and related vulnerability exposure.

As a consequences, the curl command is not in the image anymore.

The only identified impact for the moment are in docker-compose usage where curl is used for healthcheck.

To prevent this and any other issue, we need to re-integrate curl in the base image.